move error to separate file
